player 1's tiles can be seen on frontend
