+++ backend/server/server.py
@@ -1,5 +1,6 @@
 from flask import Flask, jsonify, request
 from flask_cors import CORS
+from game import Game
 
 app = Flask(__name__)
 cors = CORS(app, origins='*')
@@ -7,9 +8,15 @@ cors = CORS(app, origins='*')
 #Gamemode Page
 @app.route("/api/game_state")
 def game_state():
-    exposed = ["1B", "2B", "3B"]
-    hand    = ["RED", "RED", "GREEN"]
-    return jsonify({"exposed": exposed, "hand": hand})
+    game = Game()
+    game.deal_tiles()
+    player_1_info = game.get_player_info(1)
+    
+    return jsonify({
+        "bonus": player_1_info.get("bonus_tiles", []),
+        "exposed": player_1_info.get("exposed_hand", []),
+        "hand": player_1_info.get("hand", [])
+    })
 
 @app.route("/api/select_tile", methods=["POST"])
 def select_tile():
