+++ frontend/src/index.css
@@ -48,6 +48,11 @@ button:focus-visible {
   outline: 4px auto -webkit-focus-ring-color;
 }
 
+.full-screen-component {
+  height: 100vh;
+  width: 187vh;
+}
+
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
